Change the print function prototype

+++ 0x02-more_search_algorithms/search_algos.h
@@ -5,7 +5,7 @@
 #include <stdio.h>
 int jump_search(int *array, size_t size, int value);
 size_t min(size_t a, size_t b);
-void print(int a, int b);
+void print(size_t a, int b);
 int interpolation_search(int *array, size_t size, int value);
 int exponential_search(int *array, size_t size, int value);
 void print_array(int *array, int start, int end);
